Replace mcuboot flash_map by zephyr flash_map
This patch touch only zephyr builds of mcuboot.
The patch introduce usage of zephyr flas_map module instead
of mcuboot zephyr-only implementation. Unused flash_area_to_sectors
API of former flash_map was removed as well.
Size of sector-status-update-map entry is now defined thanks to the
minimum write size supported by the flash driver.

hal_flash_align calls were replaced by flash_area_align calls.

For avoid ambiguity former zephyr-only files flash_map.*
were renamed to flash_map_extended.*
