+++ boot/bootutil/src/bootutil_misc.c
@@ -25,7 +25,13 @@
 #include "sysflash/sysflash.h"
 #include "hal/hal_bsp.h"
 #include "hal/hal_flash.h"
+
+#ifdef __ZEPHYR__
+#include "flash_map/flash_map_extended.h"
+#else
 #include "flash_map/flash_map.h"
+#endif
+
 #include "os/os.h"
 #include "bootutil/image.h"
 #include "bootutil/bootutil.h"
@@ -355,8 +361,11 @@ boot_write_flag(int flag, const struct flash_area *fap)
     default:
         return BOOT_EBADARGS;
     }
-
+#ifdef __ZEPHYR__
+    align = flash_area_align(fap);
+#else
     align = hal_flash_align(fap->fa_device_id);
+#endif
     assert(align <= BOOT_MAX_ALIGN);
     memset(buf, 0xFF, BOOT_MAX_ALIGN);
     buf[0] = BOOT_FLAG_SET;
@@ -390,7 +399,11 @@ boot_write_swap_size(const struct flash_area *fap, uint32_t swap_size)
     uint8_t align;
 
     off = boot_swap_size_off(fap);
+#ifdef __ZEPHYR__
+    align = flash_area_align(fap);
+#else
     align = hal_flash_align(fap->fa_device_id);
+#endif
     assert(align <= BOOT_MAX_ALIGN);
     if (align < sizeof swap_size) {
         align = sizeof swap_size;
