+++ boot/bootutil/src/bootutil_priv.h
@@ -21,7 +21,13 @@
 #define H_BOOTUTIL_PRIV_
 
 #include "sysflash/sysflash.h"
+
+#ifdef __ZEPHYR__
+#include <flash_map/flash_map_extended.h>
+#else
 #include "flash_map/flash_map.h"
+#endif
+
 #include "bootutil/image.h"
 
 #ifdef __cplusplus
