+++ boot/bootutil/test/src/boot_test.c
@@ -27,7 +27,13 @@
 #include "sysflash/sysflash.h"
 #include "testutil/testutil.h"
 #include "hal/hal_flash.h"
+
+#ifdef __ZEPHYR__
+#include <flash_map/flash_map_extended.h>
+#else
 #include "flash_map/flash_map.h"
+#endif
+
 #include "bootutil/image.h"
 #include "bootutil/bootutil.h"
 #include "bootutil_priv.h"
